fix: use "number" instead of "double" for format
The server sends down 'condition: 1' even though the format is
listed as 'double' which at least the Dart OpenAPI generator treats
very strictly as a floating point which '1' does not parse as (it parses
as an integer).  OpenAPI does seem to have distinct types for
"integer", "number" and "double", since the server isn't always including
the dot, I guess this type should be "number".  At least that would
fix the current OpenAPI dart generator. If you believe the open api
generator is at fault, I'm open to fixing that instead.